usb-dwc: clear STUP interrupt flag
Previously the STUP interrupt flag was never cleared.
On the STM32F722 the STUP interrupt flag seems to stops the internal
statemachine of the USB core from reaching the setup finished state.
This renders the USB controller unusable.

This commit ensures the STUP interrupt flag is cleared whenever the
driver is polled.