usb: avoid modulo operation
The current code uses modulo using a divisor which is not necessarily
a power of two. This pulls in __aeabi_idivmod and other libgcc provided
functions dealing with division.

The code can be rewritten to determine if a zero-length packet is
actually required once we are transfering the last packet. This improves
runtime as well as code size.