stm32: usart: fix doxygen for h7 at least
The usart_common_fifos uses a very nice style of docs in the headers, so
inline help works in some editors, without having to have the source of
the library available as well.  However, it means that the group
definition with the name doesn't appear until later, and then the title
is ignored.  Move the description to the header definition instead.