SAM3A: Include GPIO in doxygen
This defines the doxygen modules peripheral_apis and SAM3A_defines
for the family sam/3a via doc-sam3a.h.

sam/3a/gpio.h now defines the submodule gpio_defines.

sam/common/gpio_*.h now use this module via the @addtogroup doxygen tag.

The result is that a module for GPIO appears in the doxygen module hierarchy
for the SAM3A family.
With the pattern of this commit, the remaining doxygen for SAM families
can be enabled.