JSQR injected sequences must be ascending.
Currently, the software loads the ADCx_JSQR in descending order.
The ADCx_JSQ4 is loaded first down to ADCx_JSQ1 when calling the
adc_set_injected_sequence function.

This causes problems if ALL the sequences are not used, because
the ones left unused will contain an invalid configuration 0x00
value.

This changes reverses the loading order such that if, for example,
three sequences are used, the JSQ1..JSQR3 registers will be loaded
with the coorrect channels.