usb_control: Rewrite USBD_REQ_NEXT_CALLBACK check for clarity
Since there are only the three values possible, this change is only a
boolean transformation and changes nothing in the execution. It makes
the code easier to read as it reflects the intention of the code better.

For the curious it also makes it easier to look up where the
USBD_REQ_NEXT_CALLBACK value comes to use.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>