NEEDSTEST: stm32f0: comparator: fix memory accesses
Fixes: https://github.com/libopencm3/libopencm3/issues/1289

While it had seemed a clever trick to access COMP_CSR as two 16bit
registers, apparently in the real world, this didn't behave as expected.

Switch the register definitions to 32bit and adjust all the functions to
auto select the correct portion of the register.

NEEDSTEST: this has been coded only, I personally have no code using
this, but it seems like the correct approach, and the current code is
obviously broken.

Signed-off-by: Karl Palsson <karlp@tweak.au>