FIX:ensure reserved bits are not touched
The reserved bits of the status register are _not_ rc_w0 therefore
a 1 in a reserved bit position is not allowed.
We thus need to ensure that all reserved bits are set in or flag
parameter so they are unset when we take the inverse.

Each type of timer peripheral has more or less reserved bits, so
we need to ensure that we set the correct bits, depending on the
peripheral.