stm32/f2/f4 bug:dma: add missing volatile qualifiers
Add missing `volatile` qualifiers to the register definitions for
DMA_SPAR, DMA_SM0AR and DMA_SM1AR.

Previously, these macros were defined as follows:

  #define DMA_SPAR(port, n)   (*(volatile void **)\
                               (DMA_STREAM((port), (n)) + 0x08))

So DMA_SPAR is a dereference of a pointer to a pointer to volatile void.
This means that DMA_SPAR itself lacks the volatile qualifier.

The lack of `volatile` would cause the compiler to reorder any writes to
these registers under high optimization levels, thereby causing code to
miscompile. Writes to these registers would be done after enabling the
DMA engine, at which point the registers are read-only.